requirements.txt
# HTTP requests for OpenRouter API
requests>=2.31.0

# Excel export
openpyxl>=3.1.0

# Load environment variables from .env file
python-dotenv>=1.0.0
